Fix bug for addWeight command that will affect graph command
Current addWeight command allows multiple adding of weights at the same
time which will affect the graph command.

Required to fix or the line graph will have issues.

Let's handle the error for the addWeight command when the user input a
weight of same date into the database.

Throwing a new command exception is the best way to do this.