Bug Fix: Allow empty email
Allowing for an email to be empty fixes the issue of previously being
unable to remove an email when editing.

Expected behaviour of the email field should follow that of an optional
field, in that it should be allowed to be empty within entries.

This fix will now allow users to input an empty string when editing the
email field.