Fix Ui to prevent clipping of TeamList
Clipping occurs as both the teamList panel and personList panel have reached their minimum width and cannot grow smaller, yet the mainWindow's minimum width is small than the sum of the two panel's min width.
The simple fix is to restrict the minimum width of the main window to be the sum of the minimum width of the two panels.