Fix entity selection with mouse
Previously, selecting entity with mouse will unfocus on the command box.
With the way `UiEventManager` was implemented, the `CommandBox` will
be brought into focus whenever a key is pressed. This led to `Enter`
refocusing on command box to the start of line, which somehow led to
the whole box being highlighted.

This has been fix now as every key press will send refocus to command box
if not already focused and position the caret at the end of the text only
when refocusing is done. This prevents the highlighting of the whole command box
when refocusing on command box.

Previously, selecting eneity with mouse in command box and pressing enter
creates a odd behaviour where the command box is filled with reference before
quickly being cleared again. This has been fix by creating a `EnterKey` key
binds which forces input submission when enter is pressed regardless of focus.
This allows the selection of entity made with mouse to be utilised despite
a selection with mouse unfocuses from the command box.