Fix update dr appmt list when deleting patient
When deleting a patient, the dr appmt list does not reflect the updated
appmt list (should remove all appmts that the deleted patient has
scheduled with that dr).

Deleting the appmts in the dr's list was not handled in the delete
command.

Let's update the execute method of the delete command.